Call start_migrate_req from bhyverun, and not from bhyvectl
Start migration the same as we start checkpoint or suspend.
From bhyvectl we send the migrate, checkpoint or suspend operation
request to a bhyve thread which listen to connections. We need to
start the migration process from the bhyverun because we need to
have access to the virtual machine context from the bhyverun.

Signed-off-by: Elena Mihailescu <elenamihailescu22@gmail.com>