Fix pci_device not used bug for migration
We were wrongly assuming that all pci devices that had the migration
functionality implemented were being used. When one of them is not used
(e.g., virtio-net), the migration process would fail. We fixed it by
setting the device size to 0 at the save end, and ignore restoring it
at the receiving end if the size is 0.

Add missing include to snapshot.h.

Add helper function to get pci devices from snapshot component.

Signed-off-by: Elena Mihailescu <elenamihailescu22@gmail.com>