Also mark vCPUs in an MTRAP exit as idle for checkpoints.
Most VM exit types will return to the kernel "soon" and so it is ok to
wait for them to do that and pop back out with a debug VM exit.  However,
mtrap exits might result in the vCPU thread waiting in the debug server
for an indeterminate amount of time, so handle checkpointing for mtrap
exits as well.